Move integration test builders into their own file
